requirements.txt
# Core dependencies
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0

# Web interface
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Machine Learning (optional)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0

# Configuration and utilities
pyyaml>=6.0
pathlib2>=2.3.7
dataclasses-json>=0.6.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU support
# torch-audio>=2.0.0  # Uncomment for audio processing
# torchaudio>=2.0.0  # Uncomment for audio processing
